feat: add map layer control, markers, measurement tools, and traffic control components
